\name{isAvailableFbcPlugin}
\alias{isAvailableFbcPlugin}

\title{
  Check for libSBML FBC plugin
}

\description{
  Check, if sybilSBML was built with libSBML containing the FBC plugin. 
}

\usage{
  isAvailableFbcPlugin()
}

\value{
  A single boolean value for the availability of the FBC plugin.
}

\author{
  Mayo Roettger <mayo.roettger@hhu.de>

  Maintainer: Mayo Roettger <mayo.roettger@hhu.de>
}


\keyword{ IO }
