\name{isAvailableGroupsPlugin}
\alias{isAvailableGroupsPlugin}

\title{
  Check for libSBML Groups plugin
}

\description{
  Check, if sybilSBML was built with libSBML containing the Groups plugin. 
}

\usage{
  isAvailableFbcPlugin()
}

\value{
  A single boolean value for the availability of the Groups plugin.
}

\author{
  Mayo Roettger <mayo.roettger@hhu.de>

  Maintainer: Mayo Roettger <mayo.roettger@hhu.de>
}


\keyword{ IO }
