\name{versionLibSBML}
\alias{versionLibSBML}

\title{
  Get libSBML Version
}

\description{
  Retrieve libSBML version number.
}

\usage{
  versionLibSBML()
}

\value{
  A single character string containing the dotted version of the
  libSBML version number.
}

\references{
  Bornstein, B. J., Keating, S. M., Jouraku, A., and Hucka M. (2008)
  LibSBML: An API Library for SBML. \emph{Bioinformatics} \bold{24},
  880--881.
}

\author{
  Gabriel Gelius-Dietrich <geliudie@uni-duesseldorf.de>

  Maintainer: Mayo Roettger <mayo.roettger@hhu.de>
}


\keyword{ IO }
